requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
datasets==2.14.4
dill==0.3.7
donut-python==1.0.9
filelock==3.18.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
lightning-utilities==0.14.3
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
munch==4.0.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.3.1
pillow==10.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.4
pydantic_core==2.18.4
PyMuPDF==1.24.6
PyMuPDFb==1.24.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==1.8.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
sconf==0.2.5
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==2.0.30
sympy==1.14.0
tensorboardX==2.6.4
timm==0.9.16
tokenizers==0.15.2
tomli==2.2.1
torch==2.1.0
torchmetrics==1.7.4
torchvision==0.16.0
tqdm==4.67.1
transformers==4.38.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.0.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zss==1.2.0